+++ mlir/test/QRef/VerifierTests.mlir
@@ -222,51 +222,3 @@ func.func @test_hermitian_bad_matrix_shape(%q0: !qref.bit, %matrix: tensor<20x20
     %obs = qref.hermitian(%matrix : tensor<20x20xcomplex<f64>>) %q0 : !quantum.obs
     return
 }
-
-// -----
-
-func.func @test_use_after_free() {
-    %r = qref.alloc(3) : !qref.reg<3>
-    %q = qref.get %r[0] : !qref.reg<3> -> !qref.bit
-    qref.custom "PauliX"() %q : !qref.bit
-    qref.dealloc %r : !qref.reg<3>
-
-    // expected-error@+1 {{Detected use of a qubit after deallocation}}
-    qref.custom "PauliX"() %q : !qref.bit
-    return
-}
-
-// -----
-
-func.func @test_use_after_free_single_qubit_alloc() {
-    %q = qref.alloc_qb : !qref.bit
-    qref.custom "PauliX"() %q : !qref.bit
-    qref.dealloc_qb %q : !qref.bit
-
-    // expected-error@+1 {{Detected use of a qubit after deallocation}}
-    qref.custom "PauliX"() %q : !qref.bit
-    return
-}
-
-// -----
-
-func.func @test_use_after_free_from_arg(%r: !qref.reg<3>) {
-    %q = qref.get %r[0] : !qref.reg<3> -> !qref.bit
-    qref.custom "PauliX"() %q : !qref.bit
-    qref.dealloc %r : !qref.reg<3>
-
-    // expected-error@+1 {{Detected use of a qubit after deallocation}}
-    qref.custom "PauliX"() %q : !qref.bit
-    return
-}
-
-// -----
-
-func.func @test_use_after_free_single_qubit_alloc_from_arg(%q: !qref.bit) {
-    qref.custom "PauliX"() %q : !qref.bit
-    qref.dealloc_qb %q : !qref.bit
-
-    // expected-error@+1 {{Detected use of a qubit after deallocation}}
-    qref.custom "PauliX"() %q : !qref.bit
-    return
-}
